+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
+## 1.12.2-1.1.0
+### Added
+- Added new Overburdened potion icon, courtesy of [Foreck1](https://www.curseforge.com/members/foreck1/projects)
+- Added `ru_ru.lang`, courtesy of [MetalloloM-Edition](https://github.com/MetalloloM-Edition)
+- Added new API method for adding custom item weights
+- Added new API method for adding custom capability handling
+- Added support for Shulker Box inventory weight
+
+### Changed
+- Configuration item weights now support ore dictionary strings
+
+### Fixed
+
+
+---
 
 ## 1.12.2-1.0.0
-### Added
 - Initial Release